Add files via upload
The attrition rate in the Sales department stood at 20.63%, which is significantly higher than the attrition rate in the Research & Development department, which was the lowest at 13.84%. This indicates that employees in the Sales department are leaving the company at a higher rate compared to other departments.

The attrition rate in the Sales department was the highest at 20.63%, followed by the Human Resources department at 19.05%, and the Research & Development department at 13.84%. This highlights the variation in attrition rates across different departments within the company.

To provide a quick overview of the attrition rates across different departments, the Sales department had an attrition rate of 20.63%, followed by the Research & Development department with 13.84%, and the Human Resources department with 19.05%.

The job role of Laboratory Technician accounted for 26.16% of the total attrition count. This indicates that employees in the Laboratory Technician role are leaving the company at a relatively higher rate compared to employees in other job roles.